// ecis-nodejs-yseq-feat_soap1218/service/ecissdk/model/data.ts
export interface DevCreateFilesUploadParams {
  body: {
    userid: number
    name: string
    size: number
    parent_path: Array<string>
  }
}

export interface DevCreateFileParams{
  body: {
    userid: number
    name: string
    sha1: string
    size: number
    groupid: number
    parentid: number
    parent_path: Array<string>
    etag: string
    must_create: boolean
  }
}

export interface Fileinfo {
  fileid: number
  userid: number
  groupid: number
  parent: number
  fname: string
  fsize: number
  ftype: string
  ctime: number
  mtime: number
  fver: number
}

export interface DevCreateFileResp {
  result: string
  msg?: string
  data?: {
    fileinfo: Fileinfo
    result: string
  }
}

export interface DevDownloadFileInfo {
  url: string
  sha1: string
  store: string
  real_store: string
  static_url: string
}

export interface DevDownloadFileResp {
  result: string
  msg?: string
  data?: {
    fileinfo: DevDownloadFileInfo
  }
}

export interface DevDownloadFileParams {
  query: {
    fileid: number
    userid: number
    req_by_internal: boolean
    x_origin: string
  }
}